// archive/packages/dataforge-old/src/generated/client-entities.ts
export interface Tag {
  id: any;
  createdAt: any;
  updatedAt: any;
  clientId?: any;
  name: any;
  slug: any;
  color: any;
  icon?: any;
  variant: any;
  sortOrder: any;
  isActive: any;
  usageCount: any;
  lastUsedAt?: any;
  metadata: any;
  tagSet: any;
  parent?: any;
  children: any;
  tasks: any;
}
// ...
export class Tag implements Tag {}